% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector_index.R
\name{calculate_sector_index}
\alias{calculate_sector_index}
\title{Calculate sector index}
\usage{
calculate_sector_index(sector_df)
}
\arguments{
\item{sector_df}{A data frame with Date and stock prices by company.}
}
\value{
A data frame with Date and sector_index columns.
}
\description{
Computes the average daily price across all companies in a sector.
}
