% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{df_cleaned2}
\alias{df_cleaned2}
\title{Cleaned S&P 500 Dataset}
\format{
A data frame with N rows and M columns.
}
\source{
Extracted from S&P 500 Excel dataset
}
\usage{
df_cleaned2
}
\description{
A preprocessed version of the S&P 500 dataset used for examples.
}
\keyword{datasets}
