% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_lstm.R
\name{moving_avg_forecast}
\alias{moving_avg_forecast}
\title{Internal: Moving Average Forecast}
\usage{
moving_avg_forecast(last_values, forecast_periods)
}
\description{
Internal: Moving Average Forecast
}
\keyword{internal}
