% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{stocksr}
\alias{stocksr}
\title{stocksr Example Data}
\format{
A list or data frame depending on its content.
}
\source{
Generated within the package.
}
\usage{
stocksr
}
\description{
Example data object used to demonstrate modeling pipeline.
}
\keyword{datasets}
