% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_lstm.R
\name{try_lstm_forecast}
\alias{try_lstm_forecast}
\title{Internal: LSTM Rolling Forecast}
\usage{
try_lstm_forecast(
  model,
  data,
  time_steps,
  forecast_periods,
  target_mean,
  target_sd
)
}
\description{
Internal: LSTM Rolling Forecast
}
\keyword{internal}
