% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_and_preprocess_data}
\alias{load_and_preprocess_data}
\title{Load and preprocess S&P 500 data}
\usage{
load_and_preprocess_data(file_path)
}
\arguments{
\item{file_path}{Path to the Excel file.}
}
\value{
A named list of data frames, each for a sector.
}
\description{
Loads Excel file of stock prices and groups companies by sector.
}
